+++ src/main/java/kr/codesqaud/cafe/domain/article/Article.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.domain.question;
+package kr.codesqaud.cafe.domain.article;
 
 
 public class Article {
