+++ src/main/java/kr/codesqaud/cafe/domain/article/repository/ArticleMemoryImpl.java
@@ -1,5 +1,6 @@
-package kr.codesqaud.cafe.domain.question;
+package kr.codesqaud.cafe.domain.article.repository;
 
+import kr.codesqaud.cafe.domain.article.Article;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
