+++ src/main/java/kr/codesqaud/cafe/domain/article/repository/ArticleRepository.java
@@ -1,4 +1,6 @@
-package kr.codesqaud.cafe.domain.question;
+package kr.codesqaud.cafe.domain.article.repository;
+
+import kr.codesqaud.cafe.domain.article.Article;
 
 import java.util.List;
