+++ src/main/java/kr/codesqaud/cafe/service/memory/UserMemoryService.java
@@ -1,12 +1,11 @@
 package kr.codesqaud.cafe.service.memory;
 
-import kr.codesqaud.cafe.domain.member.repository.UserRepository;
+import kr.codesqaud.cafe.domain.user.repository.UserRepository;
 import kr.codesqaud.cafe.dto.SignUpFormDto;
 import kr.codesqaud.cafe.dto.UpdateFormDto;
-import kr.codesqaud.cafe.domain.member.User;
+import kr.codesqaud.cafe.domain.user.User;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
